requirements.txt
Pillow==11.0.0
schedule==1.2.2
python-dateutil==2.9.0
pytesseract==0.3.13
requests==2.32.3
fuzzywuzzy==0.18.0
rapidfuzz==3.13.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0